patch init to use upstream
