+++ ol/cli/src/node/query.rs
@@ -130,3 +130,4 @@ impl Node {
         }
     }
 }
+
